Add AncestryAnalysisTable

+++ app/models/AncestryAnalysis.scala
@@ -5,7 +5,7 @@ import java.util.UUID
 case class AncestryAnalysis(
                              id: Option[Int],
                              sampleGuid: UUID,
-                             analysisMethodId: Long,
-                             populationId: Long,
+                             analysisMethodId: Int,
+                             populationId: Int,
                              probability: Double
                            )
